+++ src/service/clear-instance.ts
@@ -11,7 +11,7 @@ export async function clearInstance() {
     x.instanceId
   ).filter((x) => x !== undefined) as string[];
   if (instanceIdsByAliyun === undefined) {
-    return;
+    return [];
   }
   // 取得所有 sqliteDB 存在實體
   const instanceIdsBySqlite = sqliteDB.getInstances().map((instance) =>
@@ -22,7 +22,12 @@ export async function clearInstance() {
   const instanceIdsByAliyunOnly = instanceIdsByAliyun.filter(
     (id) => !instanceIdsBySqlite.includes(id),
   );
+  if (instanceIdsByAliyunOnly.length === 0) {
+    console.log("========= no instance to delete =========");
+    return undefined;
+  }
 
   // 刪除 instanceIdsByAliyunOnly 這些實體
   await aliyunECS.deleteInstance(instanceIdsByAliyunOnly);
+  return instanceIdsByAliyunOnly;
 }
